+++ src/renderer/features/assets/AssetsPortfolioView/ui/NetworkCard.tsx
@@ -1,15 +1,13 @@
 import { useUnit } from 'effector-react';
 
-import { BodyText, FootnoteText, Shimmering } from '@shared/ui';
-import { cnTw, totalAmount } from '@shared/lib/utils';
+import { BodyText, FootnoteText } from '@shared/ui';
+import { cnTw } from '@shared/lib/utils';
 import { AssetByChains } from '@shared/core';
-import { AssetFiatBalance } from '@entities/price/ui/AssetFiatBalance';
-import { priceProviderModel } from '@entities/price';
 import { ChainIcon } from '@entities/chain';
-import { AssetBalance, AssetLinks } from '@entities/asset';
+import { AssetLinks } from '@entities/asset';
 import { networkModel } from '@entities/network';
 import { AssetChain } from '../lib/types';
-import { AssetBalanceTooltip } from './AssetBalanceTooltip';
+import { AssembledAssetAmount } from './AssembledAssetAmount';
 
 type Props = {
   chain: AssetChain;
@@ -18,7 +16,6 @@ type Props = {
 
 export const NetworkCard = ({ chain, asset }: Props) => {
   const chains = useUnit(networkModel.$chains);
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
 
   return (
     <li role="button" tabIndex={0} className={cnTw('flex cursor-default flex-col rounded', 'transition-shadow')}>
@@ -31,19 +28,7 @@ export const NetworkCard = ({ chain, asset }: Props) => {
           </div>
         </div>
         <div className="flex flex-col items-end">
-          {chain.balance?.free ? (
-            <>
-              <AssetBalanceTooltip asset={asset} balance={chain.balance}>
-                <AssetBalance value={totalAmount(chain.balance)} asset={asset} showSymbol={false} />
-              </AssetBalanceTooltip>
-              <AssetFiatBalance amount={totalAmount(chain.balance)} asset={asset} />
-            </>
-          ) : (
-            <div className="flex flex-col gap-y-1 items-end">
-              <Shimmering width={82} height={20} />
-              {fiatFlag && <Shimmering width={56} height={18} />}
-            </div>
-          )}
+          <AssembledAssetAmount asset={asset} balance={chain.balance} />
         </div>
         <AssetLinks assetId={chain.assetId} chainId={chain.chainId} />
       </div>
