+++ src/renderer/features/assets/AssetsPortfolioView/ui/TokenBalance.tsx
@@ -1,14 +1,13 @@
 import { useUnit } from 'effector-react';
 
 import { useI18n } from '@app/providers';
-import { Icon, Tooltip, BodyText, Plate, Shimmering, FootnoteText, CaptionText } from '@shared/ui';
+import { Icon, Tooltip, BodyText, Plate, FootnoteText, CaptionText } from '@shared/ui';
 import type { AssetByChains } from '@shared/core';
-import { totalAmount } from '@shared/lib/utils';
-import { priceProviderModel, AssetFiatBalance, TokenPrice } from '@entities/price';
-import { AssetBalance, AssetIcon, AssetLinks } from '@entities/asset';
+import { TokenPrice } from '@entities/price';
+import { AssetIcon, AssetLinks } from '@entities/asset';
 import { networkModel } from '@entities/network';
 import { ChainIcon } from '@entities/chain';
-import { AssetBalanceTooltip } from './AssetBalanceTooltip';
+import { AssembledAssetAmount } from './AssembledAssetAmount';
 
 type Props = {
   asset: AssetByChains;
@@ -18,7 +17,6 @@ export const TokenBalance = ({ asset }: Props) => {
   const { t } = useI18n();
   const chain = asset.chains[0];
 
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
   const chains = useUnit(networkModel.$chains);
 
   return (
@@ -50,19 +48,7 @@ export const TokenBalance = ({ asset }: Props) => {
         className="text-text-primar text-right"
       />
       <div className="flex flex-col items-end w-[100px]">
-        {chain.balance?.free ? (
-          <>
-            <AssetBalanceTooltip asset={asset} balance={chain.balance}>
-              <AssetBalance value={totalAmount(chain.balance)} asset={asset} showSymbol={false} />
-            </AssetBalanceTooltip>
-            <AssetFiatBalance amount={totalAmount(chain.balance)} asset={asset} />
-          </>
-        ) : (
-          <div className="flex flex-col gap-y-1 items-end">
-            <Shimmering width={82} height={20} />
-            {fiatFlag && <Shimmering width={56} height={18} />}
-          </div>
-        )}
+        <AssembledAssetAmount asset={asset} balance={chain.balance} />
       </div>
       <AssetLinks assetId={asset.chains[0].assetId} chainId={chain.chainId} />
     </Plate>
